refactor: move post logic into app modules
