+++ core/services.py
@@ -1,0 +1,5 @@
+"""Temporary shim for moved services."""
+
+# TODO: remove once callers import from app-specific modules
+from posts.services import *  # noqa: F401,F403
+
